@@ -38,7 +38,6 @@
     "build": "run-s build:claer && tsc && vite build",
     "preview": "vite preview",
     "lint:fix": "eslint ./src --fix",
-    "cz": "git-cz",
     "test": "jest",
     "gitlog": "git log --pretty=format:'%Cred%h%Creset -%C(yellow)%d%Creset %s %Cblueby %an %Cgreen(%cr)%Creset'",
     "release": "standard-version --release-as",
@@ -52,17 +51,22 @@
     "react-dom": ">=18.0.0 <20.0.0"
   },
   "devDependencies": {
+    "@stylistic/eslint-plugin-ts": "^3.0.1",
     "@swc/jest": "^0.2.27",
     "@testing-library/dom": "10.3.1",
     "@testing-library/react": "16.0.0",
     "@types/jest": "^29.5.0",
-    "@types/node": "^18.15.11",
+    "@types/node": "20.17.32",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
+    "@typescript-eslint/eslint-plugin": "^8.23.0",
+    "@typescript-eslint/parser": "^8.23.0",
     "@vitejs/plugin-react-swc": "^3.0.0",
-    "cz-git": "^1.6.1",
-    "eslint": "^8.37.0",
-    "eslint-config-react-app": "^7.0.1",
+    "cz-conventional-changelog": "3.3.0",
+    "cz-customizable": "6.6.0",
+    "eslint": "^9.19.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "npm-run-all": "^4.1.5",
@@ -70,10 +74,11 @@
     "react-dom": "^19.1.0",
     "rollup-plugin-visualizer": "^5.9.0",
     "standard-version": "^9.5.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^4.9.3",
-    "vite": "^4.2.0",
-    "vite-plugin-dts": "^2.1.0",
+    "ts-jest": "^29.3.2",
+    "ts-node": "10.8.1",
+    "tsconfig-paths": "^4.2.0",
+    "vite": "^6.0.11",
+    "vite-plugin-dts": "^4.5.3",
     "vite-plugin-eslint": "^1.8.1"
   },
   "config": {
